+++ apps/c/CloverLeaf/MPI_inline/calc_dt_kernel_mpiinline_kernel.cpp
@@ -125,70 +125,125 @@ void ops_par_loop_calc_dt_kernel(char const *name, ops_block block, int dim, int
 
 
   //set up initial pointers and exchange halos if necessary
-  int base0 = args[0].dat->base_offset + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) * start[0] * args[0].stencil->stride[0];
-  base0 = base0+ (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    start[1] * args[0].stencil->stride[1];
+  long long int base0 =
+      args[0].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[0].dat->type_size
+                                                     : args[0].dat->elem_size) *
+                                     start[0] * args[0].stencil->stride[0];
+  base0 = base0 +
+          (long long int)(block->instance->OPS_soa ? args[0].dat->type_size
+                                                   : args[0].dat->elem_size) *
+              args[0].dat->size[0] * start[1] * args[0].stencil->stride[1];
   double *p_a0 = (double *)(args[0].data + base0);
 
-  int base1 = args[1].dat->base_offset + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) * start[0] * args[1].stencil->stride[0];
-  base1 = base1+ (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    start[1] * args[1].stencil->stride[1];
+  long long int base1 =
+      args[1].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[1].dat->type_size
+                                                     : args[1].dat->elem_size) *
+                                     start[0] * args[1].stencil->stride[0];
+  base1 = base1 +
+          (long long int)(block->instance->OPS_soa ? args[1].dat->type_size
+                                                   : args[1].dat->elem_size) *
+              args[1].dat->size[0] * start[1] * args[1].stencil->stride[1];
   double *p_a1 = (double *)(args[1].data + base1);
 
-  int base2 = args[2].dat->base_offset + (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) * start[0] * args[2].stencil->stride[0];
-  base2 = base2+ (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) *
-    args[2].dat->size[0] *
-    start[1] * args[2].stencil->stride[1];
+  long long int base2 =
+      args[2].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[2].dat->type_size
+                                                     : args[2].dat->elem_size) *
+                                     start[0] * args[2].stencil->stride[0];
+  base2 = base2 +
+          (long long int)(block->instance->OPS_soa ? args[2].dat->type_size
+                                                   : args[2].dat->elem_size) *
+              args[2].dat->size[0] * start[1] * args[2].stencil->stride[1];
   double *p_a2 = (double *)(args[2].data + base2);
 
-  int base3 = args[3].dat->base_offset + (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) * start[0] * args[3].stencil->stride[0];
-  base3 = base3+ (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) *
-    args[3].dat->size[0] *
-    start[1] * args[3].stencil->stride[1];
+  long long int base3 =
+      args[3].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[3].dat->type_size
+                                                     : args[3].dat->elem_size) *
+                                     start[0] * args[3].stencil->stride[0];
+  base3 = base3 +
+          (long long int)(block->instance->OPS_soa ? args[3].dat->type_size
+                                                   : args[3].dat->elem_size) *
+              args[3].dat->size[0] * start[1] * args[3].stencil->stride[1];
   double *p_a3 = (double *)(args[3].data + base3);
 
-  int base4 = args[4].dat->base_offset + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) * start[0] * args[4].stencil->stride[0];
-  base4 = base4+ (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    start[1] * args[4].stencil->stride[1];
+  long long int base4 =
+      args[4].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[4].dat->type_size
+                                                     : args[4].dat->elem_size) *
+                                     start[0] * args[4].stencil->stride[0];
+  base4 = base4 +
+          (long long int)(block->instance->OPS_soa ? args[4].dat->type_size
+                                                   : args[4].dat->elem_size) *
+              args[4].dat->size[0] * start[1] * args[4].stencil->stride[1];
   double *p_a4 = (double *)(args[4].data + base4);
 
-  int base5 = args[5].dat->base_offset + (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) * start[0] * args[5].stencil->stride[0];
-  base5 = base5+ (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) *
-    args[5].dat->size[0] *
-    start[1] * args[5].stencil->stride[1];
+  long long int base5 =
+      args[5].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[5].dat->type_size
+                                                     : args[5].dat->elem_size) *
+                                     start[0] * args[5].stencil->stride[0];
+  base5 = base5 +
+          (long long int)(block->instance->OPS_soa ? args[5].dat->type_size
+                                                   : args[5].dat->elem_size) *
+              args[5].dat->size[0] * start[1] * args[5].stencil->stride[1];
   double *p_a5 = (double *)(args[5].data + base5);
 
-  int base6 = args[6].dat->base_offset + (block->instance->OPS_soa ? args[6].dat->type_size : args[6].dat->elem_size) * start[0] * args[6].stencil->stride[0];
-  base6 = base6+ (block->instance->OPS_soa ? args[6].dat->type_size : args[6].dat->elem_size) *
-    args[6].dat->size[0] *
-    start[1] * args[6].stencil->stride[1];
+  long long int base6 =
+      args[6].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[6].dat->type_size
+                                                     : args[6].dat->elem_size) *
+                                     start[0] * args[6].stencil->stride[0];
+  base6 = base6 +
+          (long long int)(block->instance->OPS_soa ? args[6].dat->type_size
+                                                   : args[6].dat->elem_size) *
+              args[6].dat->size[0] * start[1] * args[6].stencil->stride[1];
   double *p_a6 = (double *)(args[6].data + base6);
 
-  int base7 = args[7].dat->base_offset + (block->instance->OPS_soa ? args[7].dat->type_size : args[7].dat->elem_size) * start[0] * args[7].stencil->stride[0];
-  base7 = base7+ (block->instance->OPS_soa ? args[7].dat->type_size : args[7].dat->elem_size) *
-    args[7].dat->size[0] *
-    start[1] * args[7].stencil->stride[1];
+  long long int base7 =
+      args[7].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[7].dat->type_size
+                                                     : args[7].dat->elem_size) *
+                                     start[0] * args[7].stencil->stride[0];
+  base7 = base7 +
+          (long long int)(block->instance->OPS_soa ? args[7].dat->type_size
+                                                   : args[7].dat->elem_size) *
+              args[7].dat->size[0] * start[1] * args[7].stencil->stride[1];
   double *p_a7 = (double *)(args[7].data + base7);
 
-  int base8 = args[8].dat->base_offset + (block->instance->OPS_soa ? args[8].dat->type_size : args[8].dat->elem_size) * start[0] * args[8].stencil->stride[0];
-  base8 = base8+ (block->instance->OPS_soa ? args[8].dat->type_size : args[8].dat->elem_size) *
-    args[8].dat->size[0] *
-    start[1] * args[8].stencil->stride[1];
+  long long int base8 =
+      args[8].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[8].dat->type_size
+                                                     : args[8].dat->elem_size) *
+                                     start[0] * args[8].stencil->stride[0];
+  base8 = base8 +
+          (long long int)(block->instance->OPS_soa ? args[8].dat->type_size
+                                                   : args[8].dat->elem_size) *
+              args[8].dat->size[0] * start[1] * args[8].stencil->stride[1];
   double *p_a8 = (double *)(args[8].data + base8);
 
-  int base9 = args[9].dat->base_offset + (block->instance->OPS_soa ? args[9].dat->type_size : args[9].dat->elem_size) * start[0] * args[9].stencil->stride[0];
-  base9 = base9+ (block->instance->OPS_soa ? args[9].dat->type_size : args[9].dat->elem_size) *
-    args[9].dat->size[0] *
-    start[1] * args[9].stencil->stride[1];
+  long long int base9 =
+      args[9].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[9].dat->type_size
+                                                     : args[9].dat->elem_size) *
+                                     start[0] * args[9].stencil->stride[0];
+  base9 = base9 +
+          (long long int)(block->instance->OPS_soa ? args[9].dat->type_size
+                                                   : args[9].dat->elem_size) *
+              args[9].dat->size[0] * start[1] * args[9].stencil->stride[1];
   double *p_a9 = (double *)(args[9].data + base9);
 
-  int base10 = args[10].dat->base_offset + (block->instance->OPS_soa ? args[10].dat->type_size : args[10].dat->elem_size) * start[0] * args[10].stencil->stride[0];
-  base10 = base10+ (block->instance->OPS_soa ? args[10].dat->type_size : args[10].dat->elem_size) *
-    args[10].dat->size[0] *
-    start[1] * args[10].stencil->stride[1];
+  long long int base10 =
+      args[10].dat->base_offset +
+      (long long int)(block->instance->OPS_soa ? args[10].dat->type_size
+                                               : args[10].dat->elem_size) *
+          start[0] * args[10].stencil->stride[0];
+  base10 = base10 +
+           (long long int)(block->instance->OPS_soa ? args[10].dat->type_size
+                                                    : args[10].dat->elem_size) *
+               args[10].dat->size[0] * start[1] * args[10].stencil->stride[1];
   double *p_a10 = (double *)(args[10].data + base10);
